Rename markdown-extentions file
