+++ lib/src/index.ts
@@ -2,11 +2,11 @@ export * from './clipboard-button.component';
 export * from './clipboard-options';
 export * from './katex-options';
 export * from './language.pipe';
-export * from './markdown-extensions';
 export * from './markdown.component';
 export * from './markdown.module';
 export * from './markdown.pipe';
 export * from './markdown.service';
+export * from './marked-extensions';
 export * from './marked-options';
 export * from './marked-renderer';
 export * from './mermaid-options';
