+++ lib/src/markdown.service.ts
@@ -19,8 +19,8 @@ import { map } from 'rxjs/operators';
 import { ClipboardButtonComponent } from './clipboard-button.component';
 import { ClipboardOptions, ClipboardRenderOptions } from './clipboard-options';
 import { KatexOptions } from './katex-options';
-import { MARKED_EXTENSIONS } from './markdown-extensions';
 import { ɵMARKED } from './marked';
+import { MARKED_EXTENSIONS } from './marked-extensions';
 import { MarkedOptions } from './marked-options';
 import { MarkedRenderer } from './marked-renderer';
 import { MermaidAPI } from './mermaid-options';
